+++ src/PluginEditor.h
